move metadata server files into pkg/metadata/server
The API service's server-side files are in pkg/api/server. This patch
simply moves the metadata service's server-side files into
pkg/metadata/server and out of pkg/metadata.
