+++ pkg/metadata/server/errors.go
@@ -1,4 +1,4 @@
-package metadata
+package server
 
 import (
 	"google.golang.org/grpc/codes"
