+++ pkg/metadata/server/object.go
@@ -1,4 +1,4 @@
-package metadata
+package server
 
 import (
 	"context"
