+++ pkg/metadata/server/object_type.go
@@ -1,4 +1,4 @@
-package metadata
+package server
 
 import (
 	"context"
