+++ pkg/metadata/server/property_definition_filter.go
@@ -1,4 +1,4 @@
-package metadata
+package server
 
 import (
 	"github.com/runmachine-io/runmachine/pkg/errors"
