+++ pkg/metadata/server/storage/object.go
